fix: fix burst-kernel compilation errors
- BattleRecorder.kt: fix broken imports (kotlinx.coroutines.flow/Dispatchers.X were wrong); use Flow/MutableSharedFlow/SharedFlow/asSharedFlow/flow/Dispatchers/runBlocking properly
- BurstEnhancementOrchestrator.kt: add explicit imports for risk.RiskAssessment, risk.TaskContext so the risk./battle./circuit. qualified references resolve
- EnhancedRateLimiter.kt: rename shadowing local 'val current' to 'existing' to fix conflicting declaration with parameter
- ProgressTracker.kt: make emitEvent non-suspend using tryEmit so callers (startTask/startStep/etc.) compile
- KernelDIContainer.kt: change ComponentInfo.factory type to (() -> Any?)? so nullable T is accepted
- EnhancedSkillEventBus.kt: move misplaced 'import kotlinx.coroutines.flow.map' from end of file to top
- KernelState.kt: convert duplicate local enum to typealias of com.apex.agent.domain.model.KernelState to resolve cross-package enum conflicts (BurstKernel/BurstKernelService)
- BurstKernel.kt: drop now-redundant KernelState import (resolved via local typealias)
- BurstAdaptiveGovernor.kt: merge duplicate companion objects into one (also moves DEFAULT_MAX_CONCURRENCY so ExecutionStrategy default compiles)
- QuotaManager.kt: change private val exceedAction to var so setExceedAction compiles
- PipelineCheckpoint.kt: fix unresolved 'last.key' (last is the value not the entry) -> use last.stepIndex
- FallbackChainBuilder.kt: add parens to fix operator precedence 'simpleName ?: "Unknown" to 1'
- ExecutionContextManager.kt: fix Long? - Long? subtraction (maxOfOrNull/minOfOrNull return nullable)
- EnhancedTimeoutManager.kt: fix p95 elvis type mismatch (Long? ?: Double) -> use avg.toLong()
- EnhancedToolExecutor.kt: add explicit imports for coroutineScope/async/awaitAll instead of fully-qualified calls

+++ core/burst-kernel/src/main/java/com/apex/agent/kernel/burst/enhanced/timeout/EnhancedTimeoutManager.kt
@@ -174,7 +174,7 @@ class EnhancedTimeoutManager {
         if (history.size < 5) return baseTimeout
 
         val avg = history.average()
-        val p95 = history.sorted().drop(history.size * 95 / 100).firstOrNull() ?: avg
+        val p95 = history.sorted().drop(history.size * 95 / 100).firstOrNull() ?: avg.toLong()
 
         // 超时 = P95 * 1.5，但不超过 maxTimeout
         val adapted = (p95 * 1.5).toLong()
